% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_pdf_value.R
\name{normal_pdf_value}
\alias{normal_pdf_value}
\title{Calculate PDF value of any noraml distributions anywhere}
\usage{
normal_pdf_value(mu, sigma, x)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{x}{The location at which you want to evaluate the x-value}
}
\value{
A real number
}
\description{
Calculate PDF value of any noraml distributions anywhere
}
\examples{
normal_pdf_value(0, 1, 0)
}
